Fixed some broken stuff for presentation

+++ src/AssetManagement/Animation.cpp
@@ -42,8 +42,8 @@ DrawRect<float> Animation::GetFrameSrcRect(int animFrame) const
 
   int x = (_startIdx + frame) % spriteSheet.columns;
   int y = (_startIdx + frame) / spriteSheet.columns;
-  Vector2<float> pos(x * spriteSheet.frameSize.x, y * spriteSheet.frameSize.y);
-  return DrawRect<float>(pos.x, pos.y, spriteSheet.frameSize.x, spriteSheet.frameSize.y );
+  Vector2<float> pos(static_cast<float>(x * spriteSheet.frameSize.x), static_cast<float>(y * spriteSheet.frameSize.y));
+  return DrawRect<float>(pos.x, pos.y, static_cast<float>(spriteSheet.frameSize.x), static_cast<float>(spriteSheet.frameSize.y) );
 }
 
 //______________________________________________________________________________
